feature: flash attention
在 `Janus.Cpp.cpp` 中，删除了 `main1` 函数，添加了 `do_generate` 函数以简化图像生成流程。`main` 函数现在调用 `do_generate` 并设置 `edit_text` 的初始值。`main4` 函数的实现也进行了修改，采用新的 CPU 后端初始化方式并增强了模型处理能力。

在 `language_model.cpp` 中，`LlamaDecoderLayer::build_llama_layer` 和 `run_layer` 函数进行了重大优化，增加了对张量处理和缓存机制的支持，提升了注意力机制的性能。`LayerServer::LoadRange` 和 `LanguageModel::LoadFromBin` 函数的实现也进行了调整，以确保类型安全和增强代码健壮性。

此外，在 `language_model.h` 中，`max_cached_length` 的值从2048减少到1024，并增加了 `use_flash_attn` 的静态布尔值，以支持闪存注意力机制。最后，`tensor_utils.h` 中的 `view_tensor` 函数确保维度参数正确转换为 `size_t` 类型，增强了类型安全性。

+++ Janus.Cpp/tensor_utils.h
@@ -40,7 +40,7 @@ inline ggml_tensor* view_tensor(ggml_context* ctx, ggml_tensor* tensor, DimTypes
 	constexpr size_t num_dims = sizeof...(DimTypes);
 	static_assert(num_dims <= GGML_MAX_DIMS);
 	auto type = tensor->type;
-	std::array<size_t, num_dims> ne = { dims... };
+	std::array<size_t, num_dims> ne = { static_cast<size_t>(dims)... };
 	if constexpr (num_dims == 1)
 		return ggml_view_1d(ctx, tensor, ne[0], 0);
 	else if constexpr (num_dims == 2)
